stream correct type

+++ bsl/stream_player/stream_player.py
@@ -419,7 +419,7 @@ class _Streamer:
         sinfo = StreamInfo(
             stream_name,
             n_channels=channel_count,
-            dtype="float32",
+            dtype="float64",
             sfreq=nominal_srate,
             stype="EEG",
             source_id=stream_name,
